// src/utils/enums/Login.enum.ts
export enum MenuPathEnum {
  SETTINGS = 'CONFIG',
  SETTINGS_HOME = 'Configuraciones',
  LOGOUT = 'Salir',
  LOGIN = 'Login',
  JORNADA = 'JORNADAS',
  JORNADA_HOME = 'Inicio',
  TAREA_HOME = 'Movimientos',
  TAREA = 'Movimiento',
  TAREA_CREATE = 'CrearTarea',
  TAREA_EDIT = 'EditarTarea',
  TAREA_DETALLE = 'DetalleTarea',
  TAREA_ATTEND = 'Tarea',
  CLIENTE_HOME = 'Productos',
  CLIENTE = 'ClienteList',
  CLIENTE_DETAIL = 'ClienteDetalle',
  CLIENTE_CREATE = 'ClienteCrear',
}