// src/utils/enums/IGeneral.tsx
export enum RolesType {
  ADMIN = 'ADM',
  GERENTE = 'GER',
  JEFE = 'JEF',
  OFICIAL = 'OFI'
}


export enum EstadoTareaEnum {
  FINALIZADO = 'F',
  ACTIVO = 'A',
  PROCESO = 'P',
  ELIMINADO = 'D',
}

export enum EstadoTareaKeyEnum {
  F = 'FINALIZADO',
  A = 'ACTIVO',
  P = 'PROCESO',
  D = 'ELIMINADO',
}